refactored some functions/code
